updated files in view for edit_inventory_item, static files and website_layout. added edititem() functionality in inventory controller
Static files added and website_layout updated to work with bootstrap.
edititem() functionality added, was able to test the GET via the
test_controller successfully but havent been able to test the POST yet,
may need to add in check to determine if the pk exists display this view
else display something else.